Globals and extentions

+++ Solutions/Advent.cs
@@ -39,6 +39,13 @@ static class Advent
         }
         else
         {
+            Console.WriteLine("Submit? [y|n]");
+            var confirm = Console.ReadLine()!;
+            if (confirm.ToLower() != "y")
+            {
+                Console.WriteLine("Skipping");
+                return;
+            }
             File.WriteAllText(path, answer);
         }
